+++ docs/CHANGELOG.md
@@ -1,3 +1,5 @@
 ## Changelog for v1.7.2
 
 - Use junit-jupiter-api in test scope only #160
+- Remove unnecessary logger
+- Update github actions